+++ src/components/Dashboards/Templates/VicDHDashboard.tsx
@@ -9,8 +9,6 @@ import StCounts from '../../Widgets/StCounts/StCounts';
 // TODO: Set a max card height and handle scroll voerflow
 function VicDHDashboard(props: any) {
   const {
-    setFilterList,
-    setTabValue,
     projectId,
     groupId,
   } = props;
@@ -26,8 +24,6 @@ function VicDHDashboard(props: any) {
                   <SampleSummary
                     projectId={projectId}
                     groupId={groupId}
-                    setFilterList={setFilterList}
-                    setTabValue={setTabValue}
                   />
                 </CardContent>
               </Card>
@@ -38,8 +34,6 @@ function VicDHDashboard(props: any) {
                   <StCounts
                     projectId={projectId}
                     groupId={groupId}
-                    setFilterList={setFilterList}
-                    setTabValue={setTabValue}
                   />
                 </CardContent>
               </Card>
@@ -54,8 +48,6 @@ function VicDHDashboard(props: any) {
                   <Organisations
                     projectId={projectId}
                     groupId={groupId}
-                    setFilterList={setFilterList}
-                    setTabValue={setTabValue}
                   />
                 </CardContent>
               </Card>
@@ -78,8 +70,6 @@ function VicDHDashboard(props: any) {
                   <PhessIdStatus
                     projectId={projectId}
                     groupId={groupId}
-                    setFilterList={setFilterList}
-                    setTabValue={setTabValue}
                   />
                 </CardContent>
               </Card>