+++ src/components/ProjectOverview/projTabConstants.ts
@@ -1,0 +1,32 @@
+import { TabContentProps } from '../Common/CustomTabs';
+
+export const PROJECT_OVERVIEW_TABS: TabContentProps[] = [
+  {
+    index: 0,
+    title: 'Summary',
+  },
+  {
+    index: 1,
+    title: 'Samples',
+  },
+  {
+    index: 2,
+    title: 'Trees',
+  },
+  {
+    index: 3,
+    title: 'Plots',
+  },
+  {
+    index: 4,
+    title: 'Members',
+  },
+  {
+    index: 5,
+    title: 'Proformas',
+  },
+  {
+    index: 6,
+    title: 'Datasets',
+  },
+];