+++ src/components/Widgets/PhessIdOverall/phessIdOverallSlice.tsx
@@ -2,10 +2,10 @@ import { PayloadAction, createAsyncThunk, createSlice } from '@reduxjs/toolkit';
 import { PhessIdOverall } from './phess.id.overall.interface';
 import { getUserDashboardPhessStatus } from '../../../utilities/resourceUtils';
 import LoadingState from '../../../constants/loadingState';
-import { generateDateFilterString } from '../../../utilities/helperUtils';
 import type { RootState } from '../../../app/store';
 import { ResponseObject } from '../../../types/responseObject.interface';
 import { ResponseType } from '../../../constants/responseType';
+import { generateDateFilterString } from '../../../utilities/filterUtils';
 
 interface PhessIdOverallState {
   loading: string